CheckCheque-deploy/requirements.txt
Flask==0.12.2
cloudant==2.4.0
Pillow
pytesseract
tesseract
